backend/pyproject.toml

[project]
name = "agent-mega-stack"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "openai>=1.35.0",
  "langchain>=0.2.11",
  "langchain-openai>=0.1.17",
  "langgraph>=0.2.19",
  "faiss-cpu>=1.8.0.post1",
  "langchain-community>=0.2.10",
  "minio>=7.2.7",
  "psycopg[binary]>=3.1.19",
  "redis>=5.0.7",
  "stripe>=10.7.0",
  "orjson>=3.10.7",
  "pyjwt>=2.9.0",
  "fastembed>=0.3.4"
]

[tool.uv]
index-strategy = "unsafe-best-match"
